Исправлены уязвимости и добавлена отметка когда пользователь зарегистрировался

+++ src/main/java/by/baby/blogwebsite/http/controller/LikeController.java
@@ -1,11 +1,17 @@
 package by.baby.blogwebsite.http.controller;
 
 import by.baby.blogwebsite.dto.NewLikeDto;
+import by.baby.blogwebsite.dto.UserDto;
 import by.baby.blogwebsite.service.LikeService;
+import jakarta.servlet.http.HttpSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/like")
@@ -20,15 +26,25 @@ public class LikeController {
     }
 
     @PostMapping
-    public String like(NewLikeDto newLikeDto) {
+    public String like(NewLikeDto newLikeDto,
+                       HttpSession session) {
         logger.info("Like request received: {}", newLikeDto);
+        UserDto currentUser = (UserDto) session.getAttribute("currentUser");
+        if (!Objects.equals(newLikeDto.getUserId(), currentUser.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         likeService.saveLike(newLikeDto);
         return "redirect:/blog/" + newLikeDto.getBlogId();
     }
 
     @PostMapping("/unlike")
-    public String unlike(NewLikeDto newLikeDto) {
+    public String unlike(NewLikeDto newLikeDto,
+                         HttpSession session) {
         logger.info("Unlike request received: {}", newLikeDto);
+        UserDto currentUser = (UserDto) session.getAttribute("currentUser");
+        if (!Objects.equals(newLikeDto.getUserId(), currentUser.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         likeService.deleteLike(newLikeDto);
         return "redirect:/blog/" + newLikeDto.getBlogId();
     }
